requirements.txt
openai
streamlit
langchain
langchain_community
langchain_openai
langchain_chroma
haystack-ai
